+++ src/main/Point3D.java
@@ -1,12 +1,12 @@
 package main;
 
 public class Point3D {
-    private int x;
-    private int y;
-    private int z;
+    private double x;
+    private double y;
+    private double z;
 
     //constructor
-    public Point3D(int initX, int initY, int initZ){
+    public Point3D(double initX, double initY, double initZ){
         x = initX;
         y = initY;
         z = initZ;
@@ -19,23 +19,23 @@ public class Point3D {
     }
 
     //Getters & Setters!
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
-    public void setY(int y){
+    public void setY(double y){
         this.y = y;
     }
-    public void setZ(int z) {
+    public void setZ(double z) {
         this.z = z;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
-    public int getY() {
+    public double getY() {
         return y;
     }
-    public int getZ() {
+    public double getZ() {
         return z;
     }
 }
